+++ modules/logs/r-storage.tf
@@ -1,7 +1,9 @@
 # Storage account for Logs
 module "storage_logs" {
+  count = var.logs_storage_account_creation ? 1 : 0
+
   source  = "claranet/storage-account/azurerm"
-  version = "~> 7.6.0"
+  version = "~> 7.7.0"
 
   client_name    = var.client_name
   environment    = var.environment
@@ -67,16 +69,16 @@ module "storage_logs" {
 
 # Archived Logs File Shares
 resource "azurerm_storage_share" "archivedlogs_fileshare" {
-  count                = var.logs_storage_account_enable_archived_logs_fileshare ? 1 : 0
+  count                = var.logs_storage_account_creation && var.logs_storage_account_enable_archived_logs_fileshare ? 1 : 0
   name                 = var.logs_storage_account_archived_logs_fileshare_name
-  storage_account_name = module.storage_logs.storage_account_properties.name
+  storage_account_name = one(module.storage_logs[*].storage_account_name)
   quota                = var.logs_storage_account_archived_logs_fileshare_quota
 }
 
 # Blob Archive policy
 resource "azurerm_storage_management_policy" "archive_storage" {
-  count              = lower(var.logs_storage_account_kind) == "storagev2" && var.logs_storage_account_enable_archiving ? 1 : 0
-  storage_account_id = module.storage_logs.storage_account_properties.id
+  count              = var.logs_storage_account_creation && lower(var.logs_storage_account_kind) == "storagev2" && var.logs_storage_account_enable_archiving ? 1 : 0
+  storage_account_id = one(module.storage_logs[*].storage_account_id)
 
   rule {
     name    = "Archive"