+++ modules/logs/variables.tf
@@ -46,6 +46,12 @@ variable "log_analytics_workspace_retention_in_days" {
   default     = 30
 }
 
+variable "logs_storage_account_creation" {
+  description = "Whether the dedicated Storage Account for logs is created."
+  type        = bool
+  default     = true
+}
+
 variable "logs_storage_account_kind" {
   description = "Storage Account Kind"
   type        = string